minor refactoring, to further simplify map
