+++ src/components/spatial/Map/index.tsx
@@ -1,8 +1,5 @@
-import React, { useCallback } from 'react'
+import React from 'react'
 import { MapContainer } from 'react-leaflet'
-import { useAppSelector } from '../../../state/hooks'
-import { selectFeatures } from '../../../state/geoFeaturesSlice'
-import { useDocContext } from '../../../state/DocContext'
 import { MapFeatures } from './MapFeatures'
 import { ViewportTracker } from './ViewportTracker'
 import { MapControls } from './MapControls'
@@ -12,24 +9,6 @@ interface MapProps {
 }
 
 const Map: React.FC<MapProps> = ({ children }) => {
-  const features = useAppSelector(selectFeatures)
-  const { selection, setSelection, preview } = useDocContext()
-
-  const theFeatures = preview ? preview.data.features : features
-
-  const onClickHandler = useCallback((id: string, modifier: boolean): void => {
-    if (modifier) {
-      // add/remove from selection
-      if (selection.includes(id)) {
-        setSelection(selection.filter((selectedId) => selectedId !== id))
-      } else {
-        setSelection([...selection, id])
-      }
-    } else {
-      // just select this item
-      setSelection([id])
-    }
-  }, [selection, setSelection])
 
   return (
     <MapContainer
@@ -39,10 +18,7 @@ const Map: React.FC<MapProps> = ({ children }) => {
       zoom={8}
     >
       {children}
-      <MapFeatures 
-        features={theFeatures}
-        onClickHandler={onClickHandler}
-      />
+      <MapFeatures />
       <ViewportTracker/>
       <MapControls/>
     </MapContainer>
